+++ pleroma_bot/_twitter.py
@@ -35,7 +35,7 @@ def _get_twitter_info(self):
     return
 
 
-def _get_tweets(self, version: str, tweet_id=None):
+def _get_tweets(self, version: str, tweet_id=None, start_time=None):
     """Gathers last 'max_tweets' tweets from the user and returns them
     as an dict
     :param version: Twitter API version to use to retrieve the tweets
@@ -74,57 +74,103 @@ def _get_tweets(self, version: str, tweet_id=None):
             tweets = json.loads(response.text)
             return tweets
     elif version == "v2":
-        params = {}
-        if tweet_id:
-            url = f"{self.twitter_base_url_v2}/tweets/{tweet_id}"
-        else:
-            url = f"{self.twitter_base_url_v2}/tweets/search/recent"
-            # this only gets tweets from last week
-            params.update(
-                {
-                    "max_results": self.max_tweets,
-                    "query": f"from:{self.twitter_username}",
-                }
-            )
-        # Tweet number must be between 10 and 100
-        if not (100 >= self.max_tweets > 10):
-            raise ValueError(
-                f"max_tweets must be between 10 and 100. max_tweets: "
-                f"{self.max_tweets}"
-            )
-
-        params.update(
-            {
-                "poll.fields": "duration_minutes,end_datetime,id,options,"
-                               "voting_status",
-                "media.fields": "duration_ms,height,media_key,"
-                                "preview_image_url,type,url,width,"
-                                "public_metrics",
-                "expansions": "attachments.poll_ids,"
-                              "attachments.media_keys,author_id,"
-                              "entities.mentions.username,geo.place_id,"
-                              "in_reply_to_user_id,referenced_tweets.id,"
-                              "referenced_tweets.id.author_id",
-                "tweet.fields": "attachments,author_id,"
-                                "context_annotations,conversation_id,"
-                                "created_at,entities,"
-                                "geo,id,in_reply_to_user_id,lang,"
-                                "public_metrics,"
-                                "possibly_sensitive,referenced_tweets,"
-                                "source,text,"
-                                "withheld",
-            }
+        tweets_v2 = self._get_tweets_v2(
+            tweet_id=tweet_id, start_time=start_time
         )
-        response = requests.get(
-            url, headers=self.header_twitter, params=params
-        )
-        if not response.ok:
-            response.raise_for_status()
-        tweets_v2 = json.loads(response.text)
         return tweets_v2
     else:
         raise ValueError(f"API version not supported: {version}")
 
 
-def get_tweets(self):
-    return self.tweets
+def _get_tweets_v2(
+        self, start_time, tweet_id=None, next_token=None, tweets_v2=None
+):
+    # Tweet number must be between 10 and 100
+    if not (100 >= self.max_tweets > 10):
+        raise ValueError(
+            f"max_tweets must be between 10 and 100. max_tweets: "
+            f"{self.max_tweets}"
+        )
+    params = {}
+    if tweet_id:
+        url = f"{self.twitter_base_url_v2}/tweets/{tweet_id}"
+    else:
+        url = (
+            f"{self.twitter_base_url_v2}/users/by?"
+            f"usernames={self.twitter_username}"
+        )
+        response = requests.get(
+            url, headers=self.header_twitter
+        )
+        if not response.ok:
+            response.raise_for_status()
+        response = json.loads(response.text)
+        twitter_id = response["data"][0]["id"]
+
+        url = f"{self.twitter_base_url_v2}/users/{twitter_id}/tweets"
+        if next_token:
+            params.update(
+                {
+                    "pagination_token": next_token
+                }
+            )
+
+        params.update(
+            {
+                "start_time": start_time,
+                "max_results": self.max_tweets,
+            }
+        )
+
+    params.update(
+        {
+            "poll.fields": "duration_minutes,end_datetime,id,options,"
+                           "voting_status",
+            "media.fields": "duration_ms,height,media_key,"
+                            "preview_image_url,type,url,width,"
+                            "public_metrics",
+            "expansions": "attachments.poll_ids,"
+                          "attachments.media_keys,author_id,"
+                          "entities.mentions.username,geo.place_id,"
+                          "in_reply_to_user_id,referenced_tweets.id,"
+                          "referenced_tweets.id.author_id",
+            "tweet.fields": "attachments,author_id,"
+                            "context_annotations,conversation_id,"
+                            "created_at,entities,"
+                            "geo,id,in_reply_to_user_id,lang,"
+                            "public_metrics,"
+                            "possibly_sensitive,referenced_tweets,"
+                            "source,text,"
+                            "withheld",
+        }
+    )
+
+    response = requests.get(
+        url, headers=self.header_twitter, params=params
+    )
+    if not response.ok:
+        response.raise_for_status()
+
+    if tweets_v2:
+        next_tweets = response.json()
+        for tweet in next_tweets["data"]:
+            tweets_v2["data"].append(tweet)
+        for user in next_tweets["includes"]["users"]:
+            tweets_v2["includes"]["users"].append(user)
+        for tweet_include in next_tweets["includes"]["tweets"]:
+            tweets_v2["includes"]["tweets"].append(tweet_include)
+    else:
+        tweets_v2 = json.loads(response.text)
+
+    try:
+        next_token = response.json()["meta"]["next_token"]
+        if next_token:
+            self._get_tweets_v2(tweets_v2=tweets_v2, next_token=next_token)
+    except KeyError:
+        pass
+
+    return tweets_v2
+
+
+def get_tweets(self, start_time):
+    return self._get_tweets("v2", start_time=start_time)